Add C++ -> JS -> RNBO -> JS -> C++ Latents
I've added quick and dirty support for RNBO patches that will allow us
to read latents into RNBO and send latents out of RNBO.

Essentially the RNBO patch can define an inport 'latents_in' and an
outport 'latents_out' which can be used to receive/send a list of
latent values.

Because of the way the threading works, we can't really sync up RNBO
with our background thread that is actually doing the audio processing.
Instead we just have to keep a record of what the most recent latent
values were. When we're processing the model, we send our latent
values to RNBO asynchronously and then read whatever the most recent
stored latent values are. Then at some point in the near future, RNBO
will have finished its processing and will update the current latent
values. This means that the latents are probably always behind by about
one block, if not more, but this is the best we can do for now.

It's possible that we could reduce the latency a bit if we used the C++
RNBO export and compiled that into our WASM module. That way we could
maybe run the message dispatch/receive code on the same thread as our
model processing. The js solution seems to work fine as a proof of
concept for now though.