Big refactor of app.js
This is a pretty big set of changes but well overdue. I've changed
things so that now we have this global app object that just keeps track
of all the global state. It makes it much easier to look at this object
and get a sense for the different things that are happening in the code.

One of the main changes here is moving back to initialising the audio
context in response to a user gesture. There are a multitude of annoying
reasons why we have to do this that you can read about in the comments
of app.js if you really care.

Also added a little table showing the current sensor values just like
the metrics table.