Change to TrackedPromise based initialisation
This is another fairly big change to the initialisation logic that aims
to massively simplify reasoning about the initialisation of the various
things we need on the web page.

I've introduced this idea of a TrackedPromise which is basically a
deferred promise object that allows the user to query the state of the
promise. This allows us to declare the various resources that we need
for setup to complete as await-able variables. This is very nice because
Instead of having a bunch of initialisation code all over the place that
performs some of the setup in some places and some of it in others we
can just use `await Promise.all(...)` to have a block of code that is
only executed once all the required resources are actually ready. Then
all we need to do is mark the resources as resolved when we actually
create them e.g. the audio context from a user gesture.

Since the promise state is also query-able means I've added some entries
to the metrics table that show the state of the various resources that
are required for initialisation. This we can have a bit of a debug view
showing which resources we're waiting on and which ones have been
resolved.