Sensors modularity refactor
I've changed things a bit here so that the code for each kind of sensor
is factored away into its own file. There's no real technical reason to
do this but the idea is that perhaps it's easier for the layman to look
at an implementation in a specific file.

I've also added the app.features field that allows easy
enabling/disabling of certain sensors (and in theory other features in
the future).