Use crave_enable outport instead of cRaveEnable
It's likely that composers will want to have a smooth fade in/out when
the crave processing is enabled/disabled. To accommodate this, I've
added a new outport (crave_enable) that controls whether the model
processing is bypassed or not.

This means that the patch can respond to the user changing the
cRaveEnable parameter however it likes.