Add MIDI manifest support and RNBO MIDI commands
This commit allows RNBO patches to control the playback of different
MIDI files by numeric ID.

The idea is that the user will provide a midiManifest.json file along
with their patcher in the export directory that looks something like this:

{
  "0": {
    "url": "./media/mozart.mid"
  },
  "1": {
    "url": "./media/chopin.mid"
  }
}

Then in the RNBO patch the user can make use of 3 new outports:
midi.play, midi.pause, and midi.stop. These outports take a number as an
argument which corresponds to the numeric id assigned to the entry in
the midi manifest.

Note that the './' in the url field is important because it signifies
that the url is relative to the patch directory and not an absolute URL.