Update to new model format
Luca has done some more work on crave to support a few more new models.
With these changes, the old test models we were using are deprecated.
This means you will have to re-fetch the model weights as detailed in
the readme.

Also moved to a convention of storing models in website/models to keep
things a bit tidier.